Refactor code to merge ProjectItem and Task handlers
